+++ components/admin/Dashboard.tsx
@@ -1,0 +1,12 @@
+import { checkUserIsAdmin } from "@/pages/admin";
+
+type DashboardProps = { userId: string };
+
+export default function Dashboard({ userId }: DashboardProps) {
+  return (
+    <section>
+      <h5>Dashboard {userId}</h5>
+      {checkUserIsAdmin && "yes they is"}
+    </section>
+  );
+}
