+++ lib/hooks/index.ts
@@ -1,2 +1,3 @@
 export { default as useStorage } from "./useStorage";
+export { default as useUserSession } from "./useUserSession";
 export { default as useWindowWidth } from "./useWindowWidth";
